requirements.txt
tqdm
openai
anthropic
jsonlines
python-dotenv